+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Single Drude Lorentz dispersion relation.
-* Double pole Drude Lorentz dispersion relation.
+* Single pole Drude-Lorentz dispersion relation.
+* Double pole Drude-Lorentz dispersion relation.
+* Surface plasmon polariton dispersion relation.
 
 ### Changed